Create a separate DraggableAvatar component
